refactor(Race, RaceController): 출력 관련 메소드를 RaceController로 이동
